Move `mergeRefs` to `design`
